bots/extended/pyproject.toml
[project]
name = "extended"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "structlog>=25.5.0",
    "x10-python-trading-starknet>=0.0.17",
    "pydantic-settings>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
